Codeconvention, double > long 수정
1. Double > Long 안바뀌어있던 것 수정

+++ src/main/java/service/CalculatorService.java
@@ -13,7 +13,7 @@ public class CalculatorService {
         this.calculator = calculator;
     }
 
-    public double calculate(final String formula) {
+    public long calculate(final String formula) {
         List<String> splitedFormula = Splitter.split(formula.trim());
 
         return calculator.calculate(splitedFormula);
